fixed uint problem

+++ src/can_driver.h
@@ -1,7 +1,7 @@
 #ifndef __CAN_DRIVER_H
 #define __CAN_DRIVER_H
 
-void can_send_message(int id, int length, int* arr, int transmit_line);
+void can_send_message(int id, int length, uint8_t* arr, int transmit_line);
 struct can_message can_read_message(int read_line);
